refactor(tests): Extract expandTask and expandAllTasks tests
