+++ tests/unit/task-manager/expand-all-tasks.test.js
@@ -1,0 +1,180 @@
+/**
+ * Tests for the expandAllTasks function
+ */
+import { jest } from '@jest/globals';
+import fs from 'fs';
+import path from 'path';
+
+// Mock implementations (copied from expand-task.test.js / original task-manager.test.js)
+const mockReadFileSync = jest.fn();
+const mockExistsSync = jest.fn();
+const mockMkdirSync = jest.fn();
+const mockDirname = jest.fn();
+const mockCallClaude = jest.fn().mockResolvedValue({ tasks: [] });
+const mockCallPerplexity = jest.fn().mockResolvedValue({ tasks: [] });
+const mockWriteJSON = jest.fn();
+const mockGenerateTaskFiles = jest.fn();
+const mockWriteFileSync = jest.fn();
+const mockFormatDependenciesWithStatus = jest.fn();
+const mockDisplayTaskList = jest.fn();
+const mockValidateAndFixDependencies = jest.fn();
+const mockReadJSON = jest.fn();
+const mockLog = jest.fn();
+const mockIsTaskDependentOn = jest.fn().mockReturnValue(false);
+const mockCreate = jest.fn();
+const mockChatCompletionsCreate = jest.fn();
+const mockGetAvailableAIModel = jest.fn();
+const mockPromptYesNo = jest.fn();
+const mockDisplayBanner = jest.fn();
+const mockGetTask = jest.fn();
+const mockReadComplexityReport = jest.fn(); // Added mock
+const mockFindTaskInComplexityReport = jest.fn(); // Added mock
+
+// Mock fs module
+jest.mock('fs', () => ({
+    readFileSync: mockReadFileSync,
+    existsSync: mockExistsSync,
+    mkdirSync: mockMkdirSync,
+    writeFileSync: mockWriteFileSync
+}));
+
+// Mock path module
+jest.mock('path', () => ({
+    dirname: mockDirname,
+    join: jest.fn((dir, file) => `${dir}/${file}`)
+}));
+
+// Mock ui (adjust path)
+jest.mock('../../../scripts/modules/ui.js', () => ({
+    formatDependenciesWithStatus: mockFormatDependenciesWithStatus,
+    displayBanner: mockDisplayBanner,
+    displayTaskList: mockDisplayTaskList,
+    startLoadingIndicator: jest.fn(() => ({ stop: jest.fn() })),
+    stopLoadingIndicator: jest.fn(),
+    createProgressBar: jest.fn(() => ' MOCK_PROGRESS_BAR '),
+    getStatusWithColor: jest.fn((status) => status),
+    getComplexityWithColor: jest.fn((score) => `Score: ${score}`)
+}));
+
+// Mock dependency-manager (adjust path)
+jest.mock('../../../scripts/modules/dependency-manager.js', () => ({
+    validateAndFixDependencies: mockValidateAndFixDependencies,
+    validateTaskDependencies: jest.fn()
+}));
+
+// Mock utils (adjust path)
+jest.mock('../../../scripts/modules/utils.js', () => ({
+    writeJSON: mockWriteJSON,
+    readJSON: mockReadJSON,
+    log: mockLog,
+    CONFIG: {
+        model: 'mock-claude-model',
+        maxTokens: 4000,
+        temperature: 0.7,
+        debug: false,
+        defaultSubtasks: 3
+        // Add other necessary CONFIG properties if needed
+    },
+    sanitizePrompt: jest.fn((prompt) => prompt),
+    findTaskById: jest.fn((tasks, id) =>
+        tasks.find((t) => t.id === parseInt(id))
+    ),
+    readComplexityReport: mockReadComplexityReport, // Use the actual mock
+    findTaskInComplexityReport: mockFindTaskInComplexityReport, // Use the actual mock
+    truncate: jest.fn((str, len) => str.slice(0, len)),
+    promptYesNo: mockPromptYesNo
+}));
+
+// Mock AI services (adjust path)
+jest.mock('../../../scripts/modules/ai-services.js', () => ({
+    callClaude: mockCallClaude,
+    callPerplexity: mockCallPerplexity,
+    generateSubtasks: jest.fn(),
+    generateSubtasksWithPerplexity: jest.fn(),
+    generateComplexityAnalysisPrompt: jest.fn(),
+    getAvailableAIModel: mockGetAvailableAIModel,
+    handleClaudeError: jest.fn()
+}));
+
+// Mock Anthropic SDK
+jest.mock('@anthropic-ai/sdk', () => {
+    return {
+        Anthropic: jest.fn().mockImplementation(() => ({
+            messages: {
+                create: mockCreate
+            }
+        }))
+    };
+});
+
+// Mock Perplexity using OpenAI
+jest.mock('openai', () => {
+    return {
+        default: jest.fn().mockImplementation(() => ({
+            chat: {
+                completions: {
+                    create: mockChatCompletionsCreate
+                }
+            }
+        }))
+    };
+});
+
+// Import the actual function to test (adjust path)
+import { expandAllTasks } from '../../../scripts/modules/task-manager.js';
+import { sampleTasks } from '../../fixtures/sample-tasks.js'; // Adjust path
+
+describe('expandAllTasks function', () => {
+    beforeEach(() => {
+        jest.clearAllMocks();
+        // Default mocks for successful reads/writes
+        mockReadJSON.mockResolvedValue(sampleTasks);
+        mockWriteJSON.mockResolvedValue(undefined);
+        mockReadComplexityReport.mockResolvedValue(null); // Default to no report
+    });
+
+    test('should expand all pending tasks', async () => {
+        // This test would verify that:
+        // 1. The function identifies all pending tasks
+        // 2. It expands each task with appropriate subtasks
+        // 3. It writes the updated tasks back to the file
+        expect(true).toBe(true); // Placeholder
+    });
+
+    test('should sort tasks by complexity when report is available', async () => {
+        // This test would verify that:
+        // 1. The function reads the complexity report
+        // 2. It sorts tasks by complexity score
+        // 3. It prioritizes high-complexity tasks
+        expect(true).toBe(true); // Placeholder
+    });
+
+    test('should skip tasks with existing subtasks unless force flag is set', async () => {
+        // This test would verify that:
+        // 1. The function skips tasks with existing subtasks
+        // 2. It processes them when force flag is set
+        expect(true).toBe(true); // Placeholder
+    });
+
+    test('should use task-specific parameters from complexity report', async () => {
+        // This test would verify that:
+        // 1. The function uses task-specific subtask counts
+        // 2. It uses task-specific expansion prompts
+        expect(true).toBe(true); // Placeholder
+    });
+
+    test('should handle empty tasks array', async () => {
+        // This test would verify that:
+        // 1. The function handles an empty tasks array gracefully
+        // 2. It displays an appropriate message
+        expect(true).toBe(true); // Placeholder
+    });
+
+    test('should handle errors for individual tasks without failing the entire operation', async () => {
+        // This test would verify that:
+        // 1. The function continues processing tasks even if some fail
+        // 2. It reports errors for individual tasks
+        // 3. It completes the operation for successful tasks
+        expect(true).toBe(true); // Placeholder
+    });
+});
